+++ workforus/src/main/java/site/workforus/forus/admin/controller/JobController.java
@@ -1,4 +1,4 @@
-package site.workforus.forus.job.controller;
+package site.workforus.forus.admin.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
