+++ workforus/src/main/java/site/workforus/forus/admin/service/DeptService.java
@@ -1,4 +1,4 @@
-package site.workforus.forus.dept.service;
+package site.workforus.forus.admin.service;
 
 
 import java.util.List;
@@ -7,7 +7,7 @@ import org.apache.ibatis.session.SqlSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import site.workforus.forus.dept.model.DeptDTO;
+import site.workforus.forus.admin.model.DeptDTO;
 import site.workforus.forus.mapper.DeptMapper;
 
 @Service
@@ -16,7 +16,7 @@ public class DeptService {
 	@Autowired
 	private SqlSession session;
 
-	public List<DeptDTO> selectAll() {
+	public List<DeptDTO> getAll() {
 		// Mapper 인터페이스 객체 반환
 		DeptMapper mapper = session.getMapper(DeptMapper.class);
