+++ workforus/src/main/java/site/workforus/forus/admin/service/JobService.java
@@ -1,4 +1,4 @@
-package site.workforus.forus.job.service;
+package site.workforus.forus.admin.service;
 
 import org.springframework.stereotype.Service;
